+++ scripts/cba/_helpers.py
@@ -4,6 +4,8 @@
 
 import re
 
+import pandas as pd
+
 
 def filter_projects_by_specs(
     project_list: list[str], spec_list: list[str] | str | None
@@ -84,3 +86,10 @@ def filter_projects_by_specs(
         return [p for p in project_list if p in projects]
     else:
         return [p for p in project_list if p not in projects]
+
+
+def summarize_counts(s: pd.Series):
+    ret = ""
+    for key, count in s.value_counts().items():
+        ret += f"- {key} ({count})\n"
+    return ret